Rework repo-prune
Try somewhat to keep DBs that got synced at the cut-off date working.
Instead of deleting all packages older than the given date keep one older
version of the package around, so a package install still succeeds.

In case the package was removed in the meantime and hasn't been updated
in a while we can't say if it was removed before or after the cut-off date,
so try to be conservative and keep it if it's older than 4 times the time delta.