pyproject: add extra for each CLI command
Add extras for CLI commands and move some dependencies that are only
used by them into those extras. This allows to install only a subset
of dependencies for only the command one needs (in CI for example).

To make testing easier there also exists a "all" extra which installs
everything.