sbom merge: only merge each component once
grype has started to duplicate all components in the output sbom,
which led to us assining the pkgbase link to multiple components
which ened up with duplicate CVEs in the frontend.

Avoid by only merging in the old properties once.

Alternative would be to dedupe by vuln ID in the frontend, but let's
keep it simple.