sbom/merge: add option to include the fixed versions from grype
The grype json output contains information about whether there is
a fixed version available, but doesn't include this info in the cdx
format we use.

There isn't really a direct match in cdx, but we can include the "fixed"
versions, under the "versions" field of the affected target and mark them
as "unaffected".

In the frontend we can then show the unaffected versions that are newer
than the current one as "fixed" versions.